+++ src/Stack/Setup.hs
@@ -121,8 +121,6 @@ data SetupOpts = SetupOpts
     -- ^ Do not use a custom msys installation on Windows
     , soptsResolveMissingGHC :: !(Maybe Text)
     -- ^ Message shown to user for how to resolve the missing GHC
-    , soptsSetupInfoYaml :: !FilePath
-    -- ^ Location of the main stack-setup.yaml file
     , soptsGHCBindistURL :: !(Maybe String)
     -- ^ Alternate GHC binary distribution (requires custom GHCVariant)
     , soptsGHCJSBootOpts :: [String]
@@ -225,7 +223,6 @@ setupEnv needTargets boptsCLI mResolveMissingGHC = do
             , soptsSkipGhcCheck = configSkipGHCCheck config
             , soptsSkipMsys = configSkipMsys config
             , soptsResolveMissingGHC = mResolveMissingGHC
-            , soptsSetupInfoYaml = defaultSetupInfoYaml
             , soptsGHCBindistURL = Nothing
             , soptsGHCJSBootOpts = ["--clean"]
             }
@@ -450,13 +447,13 @@ addIncludeLib (ExtraDirs _bins includes libs) config = config
 -- | Ensure both the compiler and the msys toolchain are installed and
 -- provide the PATHs to add if necessary
 ensureCompilerAndMsys
-  :: (HasConfig env, HasGHCVariant env)
+  :: (HasBuildConfig env, HasGHCVariant env)
   => SetupOpts
   -> RIO env (CompilerPaths, ExtraDirs)
 ensureCompilerAndMsys sopts = do
   didWarn <- warnUnsupportedCompiler $ getGhcVersion $ wantedToActual $ soptsWantedCompiler sopts
 
-  getSetupInfo' <- memoizeRef (getSetupInfo (soptsSetupInfoYaml sopts))
+  getSetupInfo' <- memoizeRef getSetupInfo
   (cp, ghcPaths) <- ensureCompiler sopts getSetupInfo'
 
   warnUnsupportedCompilerCabal cp didWarn
@@ -519,7 +516,7 @@ warnUnsupportedCompilerCabal cp didWarn = do
 -- | Ensure that the msys toolchain is installed if necessary and
 -- provide the PATHs to add if necessary
 ensureMsys
-  :: HasConfig env
+  :: HasBuildConfig env
   => SetupOpts
   -> Memoized SetupInfo
   -> RIO env (Maybe Tool)
@@ -549,7 +546,7 @@ ensureMsys sopts getSetupInfo' = do
       _ -> return Nothing
 
 installGhcBindist
-  :: HasConfig env
+  :: HasBuildConfig env
   => SetupOpts
   -> Memoized SetupInfo
   -> [Tool]
@@ -610,7 +607,7 @@ installGhcBindist sopts getSetupInfo' installed = do
 
 -- | Ensure compiler (ghc or ghcjs) is installed, without worrying about msys
 ensureCompiler
-  :: forall env. (HasConfig env, HasGHCVariant env)
+  :: forall env. (HasBuildConfig env, HasGHCVariant env)
   => SetupOpts
   -> Memoized SetupInfo
   -> RIO env (CompilerPaths, ExtraDirs)
@@ -652,7 +649,7 @@ ensureCompiler sopts getSetupInfo' = do
         pure (cp, paths)
 
 ensureSandboxedCompiler
-  :: HasConfig env
+  :: HasBuildConfig env
   => SetupOpts
   -> Memoized SetupInfo
   -> RIO env (CompilerPaths, ExtraDirs)
@@ -826,7 +823,7 @@ pathsFromCompiler wc compilerBuild isSandboxed compiler = withCache $ handleAny 
 buildGhcFromSource :: forall env.
    ( HasTerm env
    , HasProcessContext env
-   , HasConfig env
+   , HasBuildConfig env
    ) => Memoized SetupInfo -> [Tool] -> CompilerRepository -> Text -> Text
    -> RIO env (Tool, CompilerBuild)
 buildGhcFromSource getSetupInfo' installed (CompilerRepository url) commitId flavour = do
@@ -1070,26 +1067,25 @@ sourceSystemCompilers wanted = do
       | otherwise = id
 
 -- | Download the most recent SetupInfo
-getSetupInfo :: HasConfig env => String -> RIO env SetupInfo
-getSetupInfo stackSetupYaml = do
+getSetupInfo :: HasConfig env => RIO env SetupInfo
+getSetupInfo = do
     config <- view configL
-    setupInfos <-
-        mapM
-            loadSetupInfo
-            (SetupInfoFileOrURL stackSetupYaml :
-             configSetupInfoLocations config)
-    return (mconcat setupInfos)
+    let inlineSetupInfo = configSetupInfoInline config
+        locations' = configSetupInfoLocations config
+        locations = if null locations' then [defaultSetupInfoYaml] else locations'
+
+    resolvedSetupInfos <- mapM loadSetupInfo locations
+    return (inlineSetupInfo <> mconcat resolvedSetupInfos)
   where
-    loadSetupInfo (SetupInfoInline si) = return si
-    loadSetupInfo (SetupInfoFileOrURL urlOrFile) = do
-        bs <-
-            case parseUrlThrow urlOrFile of
-                Just req -> liftM (LBS.toStrict . getResponseBody) $ httpLbs req
-                Nothing -> liftIO $ S.readFile urlOrFile
-        WithJSONWarnings si warnings <- either throwM return (Yaml.decodeEither' bs)
-        when (urlOrFile /= defaultSetupInfoYaml) $
-            logJSONWarnings urlOrFile warnings
-        return si
+    loadSetupInfo urlOrFile = do
+      bs <-
+          case parseUrlThrow urlOrFile of
+              Just req -> liftM (LBS.toStrict . getResponseBody) $ httpLbs req
+              Nothing -> liftIO $ S.readFile urlOrFile
+      WithJSONWarnings si warnings <- either throwM return (Yaml.decodeEither' bs)
+      when (urlOrFile /= defaultSetupInfoYaml) $
+          logJSONWarnings urlOrFile warnings
+      return si
 
 getInstalledTool :: [Tool]            -- ^ already installed
                  -> PackageName       -- ^ package to find
@@ -1108,7 +1104,7 @@ getInstalledTool installed name goodVersion =
             else Nothing
     goodPackage _ = Nothing
 
-downloadAndInstallTool :: HasTerm env
+downloadAndInstallTool :: (HasTerm env, HasBuildConfig env)
                        => Path Abs Dir
                        -> DownloadInfo
                        -> Tool
@@ -1127,7 +1123,7 @@ downloadAndInstallTool programsDir downloadInfo tool installer = do
     liftIO $ ignoringAbsence (removeDirRecur tempDir)
     return tool
 
-downloadAndInstallCompiler :: (HasConfig env, HasGHCVariant env)
+downloadAndInstallCompiler :: (HasBuildConfig env, HasGHCVariant env)
                            => CompilerBuild
                            -> SetupInfo
                            -> WantedCompiler
@@ -1199,7 +1195,7 @@ getWantedCompilerInfo key versionCheck wanted toCV pairs_ =
 
 -- | Download and install the first available compiler build.
 downloadAndInstallPossibleCompilers
-    :: (HasGHCVariant env, HasConfig env)
+    :: (HasGHCVariant env, HasBuildConfig env)
     => [CompilerBuild]
     -> SetupInfo
     -> WantedCompiler
@@ -1266,41 +1262,53 @@ getOSKey platform =
         Platform AArch64               Cabal.Linux   -> return "linux-aarch64"
         Platform arch os -> throwM $ UnsupportedSetupCombo os arch
 
+downloadOrUseLocal
+    :: (HasTerm env, HasBuildConfig env)
+    => Text -> DownloadInfo -> Path Abs File -> RIO env (Path Abs File)
+downloadOrUseLocal downloadLabel downloadInfo destination =
+  case url of
+    (parseUrlThrow -> Just _) -> do
+        ensureDir (parent destination)
+        chattyDownload downloadLabel downloadInfo destination
+        return destination
+    (parseAbsFile -> Just path) -> do
+        warnOnIgnoredChecks
+        return path
+    (parseRelFile -> Just path) -> do
+        warnOnIgnoredChecks
+        root <- view projectRootL
+        return (root </> path)
+    _ ->
+        throwString $ "Error: `url` must be either an HTTP URL or a file path: " ++ url
+  where
+    url = T.unpack $ downloadInfoUrl downloadInfo
+    warnOnIgnoredChecks = do
+      let DownloadInfo{downloadInfoContentLength=contentLength, downloadInfoSha1=sha1,
+                       downloadInfoSha256=sha256} = downloadInfo
+      when (isJust contentLength) $
+        logWarn "`content-length` is not checked and should not be specified when `url` is a file path"
+      when (isJust sha1) $
+        logWarn "`sha1` is not checked and should not be specified when `url` is a file path"
+      when (isJust sha256) $
+        logWarn "`sha256` is not checked and should not be specified when `url` is a file path"
+
 downloadFromInfo
-    :: HasTerm env
+    :: (HasTerm env, HasBuildConfig env)
     => Path Abs Dir -> DownloadInfo -> Tool -> RIO env (Path Abs File, ArchiveType)
 downloadFromInfo programsDir downloadInfo tool = do
-    at <-
+    archiveType <-
         case extension of
             ".tar.xz" -> return TarXz
             ".tar.bz2" -> return TarBz2
             ".tar.gz" -> return TarGz
             ".7z.exe" -> return SevenZ
             _ -> throwString $ "Error: Unknown extension for url: " ++ url
+
     relativeFile <- parseRelFile $ toolString tool ++ extension
-    path <- case url of
-        (parseUrlThrow -> Just _) -> do
-            let path = programsDir </> relativeFile
-            ensureDir programsDir
-            chattyDownload (T.pack (toolString tool)) downloadInfo path
-            return path
-        (parseAbsFile -> Just path) -> do
-            let DownloadInfo{downloadInfoContentLength=contentLength, downloadInfoSha1=sha1,
-                             downloadInfoSha256=sha256} =
-                    downloadInfo
-            when (isJust contentLength) $
-                logWarn ("`content-length` in not checked \n" <>
-                          "and should not be specified when `url` is a file path")
-            when (isJust sha1) $
-                logWarn ("`sha1` is not checked and \n" <>
-                          "should not be specified when `url` is a file path")
-            when (isJust sha256) $
-                logWarn ("`sha256` is not checked and \n" <>
-                          "should not be specified when `url` is a file path")
-            return path
-        _ ->
-            throwString $ "Error: `url` must be either an HTTP URL or absolute file path: " ++ url
-    return (path, at)
+    let destinationPath = programsDir </> relativeFile
+    localPath <- downloadOrUseLocal (T.pack (toolString tool)) downloadInfo destinationPath
+    return (localPath, archiveType)
+
   where
     url = T.unpack $ downloadInfoUrl downloadInfo
     extension = loop url
@@ -1310,6 +1318,7 @@ downloadFromInfo programsDir downloadInfo tool = do
             | otherwise = ""
           where
             (fp', ext) = FP.splitExtension fp
+
 
 data ArchiveType
     = TarBz2
@@ -1438,7 +1447,7 @@ instance Alternative (CheckDependency env) where
             Left _ -> y
             Right x' -> return $ Right x'
 
-installGHCWindows :: HasConfig env
+installGHCWindows :: HasBuildConfig env
                   => Maybe Version
                   -> SetupInfo
                   -> Path Abs File
@@ -1451,7 +1460,7 @@ installGHCWindows mversion si archiveFile archiveType _tempDir destDir = do
     withUnpackedTarball7z "GHC" si archiveFile archiveType tarComponent destDir
     logInfo $ "GHC installed to " <> fromString (toFilePath destDir)
 
-installMsys2Windows :: HasConfig env
+installMsys2Windows :: HasBuildConfig env
                   => Text -- ^ OS Key
                   -> SetupInfo
                   -> Path Abs File
@@ -1492,7 +1501,7 @@ installMsys2Windows osKey si archiveFile archiveType _tempDir destDir = do
 
 -- | Unpack a compressed tarball using 7zip.  Expects a single directory in
 -- the unpacked results, which is renamed to the destination directory.
-withUnpackedTarball7z :: HasConfig env
+withUnpackedTarball7z :: HasBuildConfig env
                       => String -- ^ Name of tool, used in error messages
                       -> SetupInfo
                       -> Path Abs File -- ^ Path to archive file
@@ -1533,20 +1542,20 @@ expectSingleUnpackedDir archiveFile destDir = do
 -- | Download 7z as necessary, and get a function for unpacking things.
 --
 -- Returned function takes an unpack directory and archive.
-setup7z :: (HasConfig env, MonadIO m)
+setup7z :: (HasBuildConfig env, MonadIO m)
         => SetupInfo
         -> RIO env (Path Abs Dir -> Path Abs File -> m ())
 setup7z si = do
     dir <- view $ configL.to configLocalPrograms
     ensureDir dir
-    let exe = dir </> relFile7zexe
-        dll = dir </> relFile7zdll
+    let exeDestination = dir </> relFile7zexe
+        dllDestination = dir </> relFile7zdll
     case (siSevenzDll si, siSevenzExe si) of
         (Just sevenzDll, Just sevenzExe) -> do
-            chattyDownload "7z.dll" sevenzDll dll
-            chattyDownload "7z.exe" sevenzExe exe
+            _ <- downloadOrUseLocal "7z.dll" sevenzDll dllDestination
+            exePath <- downloadOrUseLocal "7z.exe" sevenzExe exeDestination
             withRunInIO $ \run -> return $ \outdir archive -> liftIO $ run $ do
-                let cmd = toFilePath exe
+                let cmd = toFilePath exePath
                     args =
                         [ "x"
                         , "-o" ++ toFilePath outdir